# riskcast-v16-main/docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: RISKCAST V3 API
  description: |
    Enterprise Insurance-Grade Risk Intelligence Platform API
    
    ## Authentication
    All endpoints require JWT authentication via Bearer token in the Authorization header.
    
    ## Versioning
    API is versioned via URL path (e.g., `/api/v3/...`)
    
    ## Rate Limiting
    - Standard: 100 requests/minute
    - Authenticated: 1000 requests/minute
    
    ## Audit Trail
    All write operations are recorded in an immutable audit ledger with hash chain verification.
    
    ## Multi-Tenancy
    All operations are automatically scoped to the authenticated user's tenant.
    
    ## Permissions
    Endpoints require specific RBAC permissions. See individual endpoint documentation.
  version: 3.0.0
  contact:
    name: API Support
    email: api-support@riskcast.com
  license:
    name: Proprietary
    url: https://riskcast.com/terms
servers:
  - url: https://api.riskcast.com/api/v3
    description: Production
  - url: https://staging-api.riskcast.com/api/v3
    description: Staging
  - url: http://localhost:8000/api/v3
    description: Local Development
...
paths:
...
  /analytics/loss-ratio:
    get:
      tags: [Analytics]
      summary: Get loss ratio metrics
      description: |
        Get loss ratio metrics for a period.
        
        **Permission Required**: `analytics:read`
      operationId: getLossRatio
      security:
        - BearerAuth: []
      parameters:
        - name: start_date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: corridor_id
          in: query
          schema:
            type: string
            format: uuid
        - name: cargo_type
          in: query
          schema:
            type: string
        - name: model_version_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Loss ratio metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LossRatioResponse'
  /analytics/model-performance/{model_version_id}:
    get:
      tags: [Analytics]
      summary: Get model performance report
      description: |
        Get performance metrics for a model version.
        
        **Permission Required**: `analytics:read`
      operationId: getModelPerformance
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ModelId'
      responses:
        '200':
          description: Model performance report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelPerformanceReport'

  # ===========================================================================
  # Compliance
  # ===========================================================================
...
components:
...
  schemas:
...
    LossRatioResponse:
      type: object
      properties:
        loss_ratio:
          type: number
        expected_loss_ratio:
          type: number
        total_premium_cents:
          type: integer
        total_loss_cents:
          type: integer
        period:
          type: object
          properties:
            start_date:
              type: string
              format: date
            end_date:
              type: string
              format: date
    ModelPerformanceReport:
      type: object
      properties:
        model_version_id:
          type: string
          format: uuid
        accuracy:
          type: number
        precision:
          type: number
        recall:
          type: number
        f1_score:
          type: number
        calibration_error:
          type: number
...
  parameters:
...
    ModelId:
      name: model_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Model version ID

  # ===========================================================================
  # Responses
  # ===========================================================================
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from authentication endpoint.
        Include in Authorization header as: `Bearer <token>`
security:
  - BearerAuth: []